requirements.txt
requests==2.31.0 
langgraph
langchain
python-dotenv==1.0.0